CRUD Transaksi perbaikan migrate dan seeder

+++ database/migrations/2024_12_26_064835_create_stoks_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->bigInteger('cabang_id')->unsigned();
             $table->bigInteger('barang_id')->unsigned();
-            $table->decimal('jumlah', 15, 2);
+            $table->integer('jumlah');
             $table->timestamps();
 
             $table->foreign('cabang_id')->references('id')->on('cabangs')->onDelete('cascade')->onUpdate('cascade');
